ON24 tap/requirements.txt
aiodns==2.0.0
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
backoff==1.8.0
cchardet==2.1.4
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
ciso8601==2.1.3
idna==2.7
jsonschema==2.6.0
mock==2.0.0
multidict==4.7.6
pbr==5.4.5
psutil==5.3.1
pycares==3.1.1
pycparser==2.20
python-dateutil==2.8.1
pytz==2018.4
requests==2.20.0
simplejson==3.11.1
singer-python==5.8.0
six==1.15.0
target-stitch==3.0.3
urllib3==1.24.3
yarl==1.4.2
